% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggReg.R
\name{AggReg}
\alias{AggReg}
\alias{Aggreg}
\title{Aggregate regions}
\usage{
Aggreg(
  sf_pol = NULL,
  n,
  nseed = 12345,
  popvec,
  areavec,
  minpop = NULL,
  maxpop = NULL,
  neigbs = NULL,
  cen_dist = NULL,
  n_reg = NULL,
  outpv = FALSE,
  nopoprist = FALSE,
  onebyone = TRUE,
  repuntilmin = FALSE
)
}
\arguments{
\item{sf_pol}{spatial polygons object}

\item{n}{the final number of regions for the aggregation}

\item{nseed}{random seed number}

\item{popvec}{population vector containing information for the starting m regional units}

\item{areavec}{area vector containing information for the starting m regional units}

\item{minpop}{minimum population of the newly created regions. If missing, then set (total population/n)*0.75.}

\item{maxpop}{maximum population of the newly created regions. If missing, then set (total population/n)*1.25.}

\item{neigbs}{neighbors list}

\item{cen_dist}{centroid distances}

\item{n_reg}{vector of indexes for the regions which will be used to choose the starting n points, default NULL, which means all regional unis from sf_pol will be used}

\item{outpv}{variables included in output. Default = FALSE, only provides a spatial polygon of the aggregated regions. If TRUE, additional variables are included (the population and area of the new regions)}

\item{nopoprist}{Logical. If there should be no population restrictions}

\item{onebyone}{Logical. I remaining regional units at the end should be added one by one}

\item{repuntilmin}{Logical. If for minimum population condition has to be fulfilled for all newly created regions}
}
\value{
\describe{\emph{newreg}} spatial polygon of the new regions (and list of neighbors \emph{nlist})
}
\description{
Given \emph{m} smaller regional units aggregate them into \emph{n} regions based on population restrictions
}
\examples{
library("RegionSim")
library("sf")
gerNUTS3<-st_read(system.file(dsn = "shape/kreise.shp", package = "RegionSim"))

simN2<-Aggreg(gerNUTS3,
              n = 38,
              popvec = gerNUTS3$popNUTS3,
              areavec = gerNUTS3$areaNUTS3,
              minpop = 678750,
              maxpop = 7000000,
              nopoprist = TRUE,
              onebyone = FALSE)
plot(simN2)

}
\references{
Simonovska R., & Tafenau E. (2023). Are the spatial scale and shape relevant in the estimation of agglomeration effects? A study on Germany
}
\seealso{
\code{\link{asgn_nbs}}
}
\author{
Rozeta Simonovska
}
