% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asgn_nbs.R
\name{asgn_nbs}
\alias{asgn_nbs}
\title{Assign neigbours}
\usage{
asgn_nbs(neigbs, cen_dist, k)
}
\arguments{
\item{neigbs}{neighbours list}

\item{cen_dist}{centroid distance matrix}

\item{k}{number of nearest neigbours to be assigned}
}
\value{
a list of neighbours
}
\description{
Assign k neigbours for regions without neighbours to a neighbours list, based on centroid distance
}
\references{
Simonovska R., & Tafenau E. (2021). Varying size and shape of spatial units: The MAUP in the case of Germany
}
\seealso{
\code{\link{AggReg}}
}
\author{
Rozeta Simonovska
}
