% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GerMunData.R
\docType{data}
\name{GerMunData}
\alias{GerMunData}
\title{German municipalities data}
\format{
a list containing:
\emph{sf_POPULATION} - vector of municipality population
\emph{sf_AREA} - vector of municipality area (in square kilometres)
\emph{n_reg} - vector of municipality ids, used as possible first pick (islands are excluded)
}
\usage{
data(GerMunData)
}
\description{
German municipalities data
}
\keyword{data}
